+++ src/us_osc_msg.c
@@ -347,11 +347,11 @@ us_osc_msg_bundle_t *
 us_osc_msg_bundle_unflatten(
     us_allocator_t *allocator,
     us_buffer_t *buf,
-    int32_t bundle_size,
+    size_t bundle_size,
     us_osc_msg_addrcode_typetag_map_proc US_UNUSED(addrmapper)
 )
 {
-    int32_t todo=bundle_size;
+    size_t todo=bundle_size;
     us_osc_msg_bundle_t *result = 0;
     us_osc_msg_bundle_t *bundle = 0;
     us_log_tracepoint();
